Improve inline suggestions
Inline suggestions were not triggering for the case of this Java
grammar:

```
pub ImportDecl: () = {
    "import" <Path> ";" => (),
    "import" <Path> "." "*" ";" => (),
};

Path: () = {
    <head: Ident> <tail: ("." <Ident>)*> => ()
};
```

This commit makes the inline suggestion a lot smarter and it now catches
this scenario; but I suspect it still needs a lot of tuning. I tried to
make it only suggest if it KNOWS it will help, except for the case of
suggesting `X?`. It may be worth doing a more intense analysis.