Allow #![] attributes to work in 0.16
In 0.16, the default mode for lalrpop will import the generated module
through a macro that `include!`s the contents.

PR #338 made it possible to use `include!` with vanilla modules by
moving all of the built-in whole-module attributes to individual
modules. However, #115 places user-generated outer attributes in the
same problematic location.

This PR moves user-supplied outer attributes to the same location as
the fix in #338.