Detect when we are building our own workspace.
This is possibly one of the most horrifying things I have ever come up
with. By putting two versions of a crate into different branches, and
using [replace] in Cargo.toml to override the dependency onto the other
branch, we can detect whether the root build is in the lalrpop
workspace.

This ensures that lalrpop does not get built more than once when used as
a dependency. This is clearly "better" than the old system.