test_util::compare: Normalize Spans in diff output, too.
`Span`s should be normalized not just when we decide whether the test has passed
or failed, but also in the strings passed to `diff::lines`. Otherwise, the
actual difference the test is unhappy about is hidden amidst dozens of `Span`
changes.

`{:#?}` includes newlines in the printed representation of `Span` values, so the
regular expression needs to be ready for that.