Fix #573
When using an external lexer with generic type bounds, those type bounds
where not propagated to the definition of __ToTriple, which this PR
fixes.

This uncovered a bug when generating the parse module as the passed-in
types where not accounting for the 2 extra type parameters for external
token and lexer.