Fail CI if any lines fail
The default bash return code is the return code of the last function
called.  This means that the CI will only mark as failed if the last
command in ci.sh fails.

This can be observed in recent CI runs (for example
https://github.com/lalrpop/lalrpop/actions/runs/4765777324/jobs/8471946793)
where tests failed, but the CI marked as passed anyways.