Handle dangling symlinks
This was a regression; we used to print a warning and otherwise ignore
dangling symlinks. This commit should restore the original behavior, and
I've tested the fix manually. I'm not sure what happens in case of a
chain of symlinks where one of the later ones (not the first) is
dangling, but I suspect we're printing the first symlink in that case.