Traverse into symlinks to dirs (#852)
* Traverse into symlinks to dirs

The lalrpop_files function traverses recursively into subdirectories of
the provided root directory and returns all .lalrpop files. However,
lalrpop_files used to skip symlinks to directories because it used the
DirEntry::file_type method to detect directories, which doesn't follow
symlinks. Since lalrpop_files does follow symlinks to files, this old
behavior for symlinks was inconsistent and unintuitive.

With this commit, we're now using the Path::is_dir method, which does
follow symlinks, so now lalrpop_files traverses also into symlinks to
directories.

* Use WalkDir to search for .lalrpop files

* Do not use same_file_system

Two issues with it:
1. It rules out some legitimate (i.e. directed, not containing loops)
   mount setups.
2. It looks like it does not work at all on unsupported OSs (i.e. not
   Windows, not Unix), so directory traversal would be completely broken
   there.

* Handle dangling symlinks

This was a regression; we used to print a warning and otherwise ignore
dangling symlinks. This commit should restore the original behavior, and
I've tested the fix manually. I'm not sure what happens in case of a
chain of symlinks where one of the later ones (not the first) is
dangling, but I suspect we're printing the first symlink in that case.