Update to itertools 0.13 (#899)
* Update to itertools 0.13

Switch deprecated group_by() to chunk_by()

The change in name was made because chunk_by() better implies that the function groups *consecutive* elements (rather than searching through the iterator and funding *all* matching elements).  This raises the question of whether the intention in the original group_by() was correct, since the change was made to avoid subtle bugs around assuming that this grouped all entries rather than consecutive entries.  In the two instances we have of this function call, the first one has a comment specifically calling out a desire to do this consecutively.  In the second, the iterator was previously sorted by the chunking value, making consecutive equivalent to all, so I believe we are free of this potential bug.

* Update lalrpop/src/lr1/codegen/parse_table.rs

Fix comment typo

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@gmail.com>

---------

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@gmail.com>