Add a new process_src() public helper to look in the source directory.
Update the docs to recommend this as the top choice for setting up
lalrpop.

Prior to this commit, the book incorrectly stated that `process_root()`
looked in `./src`, while the documentation comments in the code
correctly identified it as parsing the current directory.  Switching the
book to recommend `process_src()` addresses this inconsistency as well.