Add an example showing how the parser can control the lexer's mode.
In some languages, it is necessary for the parser to control the
lexer's mode.  This is possible, but non-obvious.

This commit adds a simple example that shows how to parse the "list of
length-value" language in which values are encoded as a length,
followed by a colon, and then the literal number of bytes, e.g.,
`2:hi`.  To parse this, the parser needs to tell the lexer to return
the next `n` literal bytes when it sees the length prefix.

See #802, #706.