Don't call exit() in a library, instead return an error to the user (#920)
This lets users who want to proceed even on lalrpop failure do so.  The
common case of calling .unwrap() on lalrpop's return will continue to
exit at (roughly the same place), seeing a panic rather than an exit,
which is probably more in line with expected behavior.

The approach in this commit is the least invasive way to achieve this
goal, since the existing returns are io::Result, this commit just
returns an io::Error (mostly ErrorKind::InvalidData) on various lalrpop
failures.  An alternative approach would be to define a lalrpop error
type and return that.  The existing lalrpop_util::ParseError covers some
but not all of the relevant cases, since it only covers errors at the
parse stage, not later stages of compilation.  I'm not sure the effort
to fully propagate all the errors up is worth the benefit, since the
error messages are displayed anyways, and I don't know what the use case
for actually parsing specific lalrpop compilation errors
programmatically from build.rs is.