Lint on calls to process::exit() (#923)
Add an exception for main.rs.  exit() calls are unexpected behavior in a
library, where the caller should have control on whether errors are
fatal or not.  In binaries, there is no caller and the arguments against
exit() are weaker.  In many cases, exit() may still be a control-flow
anti-pattern, but exceptions exist, so allow this for everything that's
definitely only in a binary context.

Regardless of the allow() in main.rs, the clippy lint contains an exception for exit calls in entrypoint
functions, so the existing exit() call, which is in fn main() does not
trigger this lint, even without the allow() attribute.