If there are both naive and better errors available for a given conflict, supress the naive ones.
When reporting an error, conflict resolution creates error states for
the all the terminals in the grammar, which can be a lot of errors.
Some of these will hopefully classify as the more helpful error
messages, but for terminals that can't shift to a valid production,
lalrpop won't be able to produce a good error message.  This can result
in error messages that have many naive errors burying the more helpful
ones.

Check what sorts of errors we will be displaying before displaying them.
If all we got is naive errors, just display those.  If we got some good
errors, use those instead and discard the naive errors.