When encountering naive errors on reduce/reduce conflicts, drop all but the first.
On conflicts, lalrpop tries every possible lookahead token to find good
conflict examples.  On shift/reduce conflicts that makes a lot of sense,
different lookahead tokens are different conflicts, because the shift is
different.  On reduce/reduce conflicts on the other hand, the lookahead
is sort of irrelevant.  Hopefully, at least one lookahead produces a
good example, so we won't need naive errors at all.  The fact that we
only found naive errors means that none of the lookaheads were helpful.
So rather than reporting one error message per token, just drop all but
the first, since we can't tell what lookahead is more helpful (if we'd
found a helpful lookahead, we'd be using that for a better error).