Add `no_std` support for `lalrpop-util` crate with `lexer` feature enabled. (#952)
This unlinks the `lexer` feature from the `std` feature. While expanded `no_std` support is possible in `lalrpop`, this is the key component that prevents using a the lexer within a `no_std` environment.

In order to do this, the `regex-automata` crate needs to have the `perf` feature removed, because it implies the `perf-literal-multisubstring` feature, which implies the `std` feature. To minimize the changes caused by this, the `perf` feature is re-enabled for the `std` feature of the `lalrpop-util` crate.