Add a new error message for ambiguous reductions to the same nonterminal (#927)
* Add a new error message for ambiguous reductions to the same nonterminal

The existing error messages assume that the source of a reduce/reduce conflict is which nonterminal is reduced to.  This is a confusing result if there are two possible reductions to the same nonterminal ("I can reduce to an A, but I can also reduce to an A").  Identical reductions to the same nonterminal are easy to do and can be hard to spot in the presence of macros.

This error message detects when the conflict is for the same nonterminal and helpfully points to the specific rules in that nonterminal that conflict.

* Address unresolved code review comments from #926

* Show line numbers in same reduction error

* Simplify span_text()