Use our README as a top level docstring
This shows on the homepage of lib.rs, and is currently blank.  The
README contains lots of good information and points users at the various
documentation sources, so make it more visible to people checking out
the crate on docs.rs.

While we're at it, warn about missing docs (which would have caught this
omission).