Specify optional regex_automata as "dep:".
https://doc.rust-lang.org/cargo/reference/features.html#optional-dependencies

I recently learned that optional dependencies will be exposed as
features unless the are explicitly marked as "dep:" for a another
feature.  This means that lalrpop-util has a feature called "regex-automata"
currently (see https://docs.rs/crate/lalrpop-util/0.22.0/features),
which isn't what we want.  We want to have the dependency optionally
when certain other features are enabled, but not expose it directly.

This is technically a breaking change
(https://doc.rust-lang.org/cargo/reference/semver.html#cargo-feature-remove).
However, it would surprise me if any users use this feature.  In the
interest of caution, it might make sense to hold this until we plan a
breaking release anyways, since it's low priority.