Stop using in_dir once we get to process_* family functions
The prior code has weird corner cases around the difference between
in_dir and the path passed to process_dir().  In many cases, they are
the same, but sometimes they are not.  When they are not, the only
behavior of in_dir is to effect where *output* files are placed, which
is weird.  Depending on the various function calls, sometimes in_dir
sets the path to search, but sometimes not, without rhyme or reason.

This commit changes in_dir to set the path to search if process() is
called specifically.  Other process_*() family functions have a built in
path to search, in which case in_dir is illegal.

Changing the output path based on the setting of in_dir is confusing and
unexpected for users.  Drop that.  That seems to have originally been
maintained for backwards compatibility.  The result of this commit is
that sometimes generated .rs files that were previously written in the
source repository will end up in the OUT_DIR instead, which is a
breaking change.  That said, OUT_DIR really should be the correct and
preferred place for generated code.  Having users explicitly opt out
from that makes more sense than opting them out surprisingly based on
the in_dir setting.

For historical discussion on this, see:
https://github.com/lalrpop/lalrpop/issues/908#issuecomment-2176200995
https://github.com/lalrpop/lalrpop/issues/280