Add some doc comments to lalrpop-util (#984)
* Add some doc comments to lalrpop-util

This isn't everything triggered by warn(missing-docs).  I tried to hit
the low hanging fruit and important stuff, but there are still quite a
few gaps, so we can't turn on the lint for it yet.

* Mark the built in lexer and the state machine as hidden in the docs

These are public for the purpose of generated parser consumption, not
end user consumption