Drop LALR support
As suggested in #759.

This makes Lane tables the only available algorithm to generate LR(1)
states.  In theory, lane tables are strictly better, and they have been
the default in lalrpop for quite some time.

Searching github reveals no public users of LALR that aren't directly lalrpop
test code.

This is of course a breaking change, since it drops support for an
existing feature.