Bump MSRV to 1.81 and remove core::error workaround
1.83 was released on 11-28, making this consistent with our MSRV policy.

This is also a sufficient MSRV for #978 and #981, which need MSRVs bumps, and lets #979 take advantage of fs::exists if desired (fs::exists was stabilized in 1.81).

Per rust's guidance, MSRV bumps are not technically breaking, but should be bundled with breaking changes if possible.  Since #979 is breaking, it makes sense to me to take both PRs at once for the 0.23 release, when we are ready for that.