Add some extra tests for untested areas of the codebase. (#1011)
* Add a test for lr1 report generation

This little-used feature didn't have any test coverage

* Convert lexer-modes main to a unit test

Currently, we're not actually running this main function to verify that
lexer-modes still works.  The "main" is really just test code, so treat
it as such.

* Convert lexer main to a unit test

Currently, we're not actually running this main function to verify that
the lexer example still works.  The "main" is really just test code, so treat
it as such.