Don't run cargo update on msrv test
This lets us not break on msrv bumps in our dependencies in keeping with
the policy at https://doc.rust-lang.org/nightly/cargo/reference/rust-version.html
which says "it is expected that each dependency’s version-requirement
supports at least one version compatible with your rust-version.
However, it is not expected that the dependency specification excludes
versions incompatible with your rust-version."

A downside is that committing an updated Cargo.lock may cause msrv
breakage at that time.