Silence new clippy lint literal_string_with_formatting_args in codegen
This new lint checks if we use format args in string literals that
aren't passed to function that handle them like println!().  This makes
sense normally, but in code generation contexts, we may be generating
code like `rust!(rust, "println!("some format arg {}", arg))` which is a
false positive hit for this lint.  There is only one such lint right
now, but just blanket silence it in codegen and leave it on for the rest
of the codebase, because it doesn't make sense in code generation.