Drop the grammar::token module
This module is unusued.  It was actually one of the very first modules created
in the very first commit of this project.  It was used in early
experiments about how to represent tokens, but was refactored out within
a few days of early lalrpop development.  It has not been included in
the lalrpop module tree or used at all since 2015.