Move the bulk of lalrpop into a new lalrpop-internals crate
This is so that we can support requests to access lalrpop internal
functions and data structures without committing to bump semver on
internal changes.

Example use cases include:
* IDE integration (#1013)
* benchmarking
* fuzzing (#924)
* program synthesis (#934)
* Documentation generation (#658)

As noted in the linked issues and PRs above, there has been interest in
this from multiple users with multiple use cases.

The scheme here is that lalrpop-internals would expose whatever is
needed for these use cases publically, and would bump semver liberally.
The main lalrpop crate only holds the library and binary front-ends, and
depends on lalrpop-internals for nearly all the work.

The public API of lalrpop-internals right now includes what is needed
for the following use cases:

* lalrpop itself
    * build, log, session
* The documentation generator in #658
    * log, grammar, parser, tok
* The LSP demo at https://github.com/LighghtEeloo/lalrpop/tree/lsp
    * build, file_text, grammar, normalize, parser, session

I haven't gone through the public members of these modules.  It's likely
that many of them should be marked pub(crate) rather than pub.  That is
something that should be done before this is merged.