Correctly put output files at the found path.
Worth noting is that this changes update_lrgrammar.sh.  The previous
setting to out_dir of "." and targeting a specific file makes little
sense to me.  Why when we supply a full path should lalrpop recreate
that path in the outdir.  If we supplied the parent and searched, then
recreating the path makes sense.  It seems to me that if the user
supplies a full path to a specific file, we should put that file
specifically in the OUT_DIR.

This includes the patch supplied by @Pat-Lafon at
https://github.com/lalrpop/lalrpop/pull/1055#issuecomment-2900061189.

Co-authored-by: Patrick LaFontaine <32135464+Pat-Lafon@users.noreply.github.com>