Handle LockResults more explicitly
If another thread panics while holding the mutex, we get an error when
unlocking it.  This is normal and expected in our case - that means a
different test failed.  Cleanup should happen via TestState.drop()
during that thread's panic.  We verify that that actually happened, and
clear the poison to clean up the mutex.