Make the lexer feature a default feature for lalrpop-util
The assumption in the code seems to have been that because lalrpop
depends on lalrpop-util, we can let lalrpop set the feature and depend
on feature unification to keep them in sync.  This simplifies things for
the end user since they only need to set the feature in one place.
However, resolver version 2 breaks this approach by keeping build
dependencies and normal dependencies from performing feature unification
with each other.  So the result is that a user who enables the lexer
feature on lalrpop, but not lalrpop-util, will get a build-dependency of
lalrpop-util with the feature enabled, but not a runtime dependency,
when they use lalrpop-util in their own code.  For this reason, it is
necessary to control these features together.

The documentation in the quick start guide reflects this assumption that
changing just lalrpop is sufficient to turn off the built in lexer.  As
of resolver version 2, that assumption is false.  Fix the documentation
to accord with the state of this PR - that the lexer is on by default in
lalrpop-util, and must be disabled in two places.

It's worth noting explicitly the user-visible aspects of this PR.  Prior
to this PR, with resolver version 2, a user using the built in lexer
would need to enable it for lalrpop-util, but could rely on the default
for lalrpop.  A user using a custom lexer would only need to disable it
in lalrpop.  After this PR, the user using the built in lexer needs no
special steps, the default features will work.  A user using a custom
lexer needs to disable the feature in two places.  This change is
desirable because 1. It shifts the burden onto users with more complex
use cases and keeps the "easy path" of using the built in lexer easy. 2.
It has consistent behavior for both libraries.