Reintroduce in dir fix (#1082)
* Revert "Revert "Stop using in_dir once we get to process_* family functions (#979)" (#1077)"

This reverts commit 21f605036a45223139707ce35dcbd02166d2b0d7.

* Clean up CUSTOM_TEST_DIR appropriately after failures

This concurrency bug wasn't seen when testing with "cargo test", but
when trying to enable cargo-nextest support, it was hitting consistently
due to the extra encapsulation provided there.

* Address uninlined-format-args lint

* Correctly put output files at the found path.

Worth noting is that this changes update_lrgrammar.sh.  The previous
setting to out_dir of "." and targeting a specific file makes little
sense to me.  Why when we supply a full path should lalrpop recreate
that path in the outdir.  If we supplied the parent and searched, then
recreating the path makes sense.  It seems to me that if the user
supplies a full path to a specific file, we should put that file
specifically in the OUT_DIR.

This includes the patch supplied by @Pat-Lafon at
https://github.com/lalrpop/lalrpop/pull/1055#issuecomment-2900061189.

Co-authored-by: Patrick LaFontaine <32135464+Pat-Lafon@users.noreply.github.com>

* Mark set_var as unsafe.

Includes a rather verbose comment about the safety concerns.  The
highlight is the recommendation to use cargo-nextest.

* Implement teardown as a drop trait

* Handle LockResults more explicitly

If another thread panics while holding the mutex, we get an error when
unlocking it.  This is normal and expected in our case - that means a
different test failed.  Cleanup should happen via TestState.drop()
during that thread's panic.  We verify that that actually happened, and
clear the poison to clean up the mutex.

---------

Co-authored-by: Patrick LaFontaine <32135464+Pat-Lafon@users.noreply.github.com>