Check for in_dir conflicts in process_dir() and process_file()
These functions should never be called with in_dir set, because they
take an input directory as an argument.  Other functions like
process_current_dir(), process_root(), process_src() call into these, so
they're in these checks as well.

We set the in_dir in process_dir() so that file paths in the output can
strip it later.  I think we missed some cases on this before because we
only did that in process_current_dir()