Mark the correct versions in deprecated attributes
As far as I can tell, the "since" field isn't actually used anywhere,
it's only there for people browsing source.  Clippy checks that it's
valid semver.

There was prior discussion on this here: https://github.com/orgs/lalrpop/discussions/823#discussioncomment-6703836

The main reasons to avoid fixing the numbers there were to avoid code
churn, and to not make things start as deprecated in a non-breaking
release.

The main things that have changed since then are that 1.0.0 seems less
imminent, and that we are gearing up for a breaking release.

So let's take the opportunity and fix these to reflect the actual
history.  That will help us make future judgments about possibly
dropping the deprecated items in a future breaking release.